+++ packaging/pep517_backend/__main__.py
@@ -1,0 +1,6 @@
+import sys
+
+from . import cli
+
+if __name__ == "__main__":
+    sys.exit(cli.run_main_program(argv=sys.argv))