+++ packaging/pep517_backend/_backend.py
@@ -14,12 +14,6 @@ from sys import version_info as _python_version_tuple
 from tempfile import TemporaryDirectory
 from warnings import warn as _warn_that
 
-try:
-    from tomllib import loads as _load_toml_from_string
-except ImportError:
-    from tomli import loads as _load_toml_from_string  # type: ignore[no-redef]
-
-from expandvars import expandvars
 from setuptools.build_meta import build_sdist as _setuptools_build_sdist
 from setuptools.build_meta import build_wheel as _setuptools_build_wheel
 from setuptools.build_meta import (
@@ -43,16 +37,19 @@ from distutils.dist import DistributionMetadata as _DistutilsDistributionMetadat
 with suppress(ImportError):
     # NOTE: Only available for wheel builds that bundle C-extensions. Declared
     # NOTE: by `get_requires_for_build_wheel()` and
-    # NOTE: `get_requires_for_build_editable()`, when `--pure-python`
+    # NOTE: `get_requires_for_build_editable()`, when `pure-python`
     # NOTE: is not passed.
     from Cython.Build.Cythonize import main as _cythonize_cli_cmd
 
 from ._compat import chdir_cm
-from ._transformers import (  # noqa: WPS436
-    get_cli_kwargs_from_config,
-    get_enabled_cli_flags_from_config,
-    sanitize_rst_roles,
+from ._cython_configuration import (  # noqa: WPS436
+    get_local_cython_config as _get_local_cython_config,
 )
+from ._cython_configuration import (
+    make_cythonize_cli_args_from_config as _make_cythonize_cli_args_from_config,
+)
+from ._cython_configuration import patched_env as _patched_cython_env
+from ._transformers import sanitize_rst_roles  # noqa: WPS436
 
 __all__ = (  # noqa: WPS410
     'build_sdist',
@@ -70,7 +67,13 @@ __all__ = (  # noqa: WPS410
 )
 
 
-PURE_PYTHON_CONFIG_SETTING = '--pure-python'
+CYTHON_TRACING_CONFIG_SETTING = 'with-cython-tracing'
+"""Config setting name toggle to include line tracing to C-exts."""
+
+CYTHON_TRACING_ENV_VAR = 'YARL_CYTHON_TRACING'
+"""Environment variable name toggle used to opt out of making C-exts."""
+
+PURE_PYTHON_CONFIG_SETTING = 'pure-python'
 """Config setting name toggle that is used to opt out of making C-exts."""
 
 PURE_PYTHON_ENV_VAR = 'YARL_NO_EXTENSIONS'
@@ -83,79 +86,55 @@ IS_CPYTHON = _system_implementation.name == "cpython"
 """A flag meaning that the current interpreter implementation is CPython."""
 
 PURE_PYTHON_MODE_CLI_FALLBACK = not IS_CPYTHON
-"""A fallback for `--pure-python` is not set."""
+"""A fallback for ``pure-python`` is not set."""
+
+
+def _is_truthy_setting_value(setting_value) -> bool:
+    truthy_values = {'', None, 'true', '1', 'on'}
+    return setting_value.lower() in truthy_values
+
+
+def _get_setting_value(
+        config_settings: dict[str, str] | None = None,
+        config_setting_name: str | None = None,
+        env_var_name: str | None = None,
+        *,
+        default: bool = False,
+) -> bool:
+    user_provided_setting_sources = (
+        (config_settings, config_setting_name, (KeyError, TypeError)),
+        (os.environ, env_var_name, KeyError),
+    )
+    for src_mapping, src_key, lookup_errors in user_provided_setting_sources:
+        if src_key is None:
+            continue
+
+        with suppress(lookup_errors):  # type: ignore[arg-type]
+            return _is_truthy_setting_value(src_mapping[src_key])  # type: ignore[index]
+
+    return default
 
 
 def _make_pure_python(config_settings: dict[str, str] | None = None) -> bool:
-    truthy_values = {'', None, 'true', '1', 'on'}
-
-    user_provided_setting_sources = (
-        (config_settings, PURE_PYTHON_CONFIG_SETTING, (KeyError, TypeError)),
-        (os.environ, PURE_PYTHON_ENV_VAR, KeyError),
+    return _get_setting_value(
+        config_settings,
+        PURE_PYTHON_CONFIG_SETTING,
+        PURE_PYTHON_ENV_VAR,
+        default=PURE_PYTHON_MODE_CLI_FALLBACK,
     )
-    for src_mapping, src_key, lookup_errors in user_provided_setting_sources:
-        with suppress(lookup_errors):  # type: ignore[arg-type]
-            candidate_val = src_mapping[src_key].lower()  # type: ignore[index]
-            return candidate_val in truthy_values
-
-    return PURE_PYTHON_MODE_CLI_FALLBACK
 
 
-def _get_local_cython_config():
-    """Grab optional build dependencies from pyproject.toml config.
-
-    :returns: config section from ``pyproject.toml``
-    :rtype: dict
-
-    This basically reads entries from::
-
-        [tool.local.cythonize]
-        # Env vars provisioned during cythonize call
-        src = ["src/**/*.pyx"]
-
-        [tool.local.cythonize.env]
-        # Env vars provisioned during cythonize call
-        LDFLAGS = "-lssh"
-
-        [tool.local.cythonize.flags]
-        # This section can contain the following booleans:
-        # * annotate — generate annotated HTML page for source files
-        # * build — build extension modules using distutils
-        # * inplace — build extension modules in place using distutils (implies -b)
-        # * force — force recompilation
-        # * quiet — be less verbose during compilation
-        # * lenient — increase Python compat by ignoring some compile time errors
-        # * keep-going — compile as much as possible, ignore compilation failures
-        annotate = false
-        build = false
-        inplace = true
-        force = true
-        quiet = false
-        lenient = false
-        keep-going = false
-
-        [tool.local.cythonize.kwargs]
-        # This section can contain args that have values:
-        # * exclude=PATTERN      exclude certain file patterns from the compilation
-        # * parallel=N    run builds in N parallel jobs (default: calculated per system)
-        exclude = "**.py"
-        parallel = 12
-
-        [tool.local.cythonize.kwargs.directives]
-        # This section can contain compiler directives
-        # NAME = "VALUE"
-
-        [tool.local.cythonize.kwargs.compile-time-env]
-        # This section can contain compile time env vars
-        # NAME = "VALUE"
-
-        [tool.local.cythonize.kwargs.options]
-        # This section can contain cythonize options
-        # NAME = "VALUE"
-    """
-    config_toml_txt = (Path.cwd().resolve() / 'pyproject.toml').read_text()
-    config_mapping = _load_toml_from_string(config_toml_txt)
-    return config_mapping['tool']['local']['cythonize']
+def _include_cython_line_tracing(
+        config_settings: dict[str, str] | None = None,
+        *,
+        default=False,
+) -> bool:
+    return _get_setting_value(
+        config_settings,
+        CYTHON_TRACING_CONFIG_SETTING,
+        CYTHON_TRACING_ENV_VAR,
+        default=default,
+    )
 
 
 @contextmanager
@@ -216,39 +195,18 @@ def patched_dist_get_long_description():
 
 
 @contextmanager
-def patched_env(env):
-    """Temporary set given env vars.
-
-    :param env: tmp env vars to set
-    :type env: dict
-
-    :yields: None
-    """
-    orig_env = os.environ.copy()
-    expanded_env = {name: expandvars(var_val) for name, var_val in env.items()}
-    os.environ.update(expanded_env)
-    if os.getenv('YARL_CYTHON_TRACING') == '1':
-        os.environ['CFLAGS'] = ' '.join((
-            os.getenv('CFLAGS', ''),
-            '-DCYTHON_TRACE=1',
-            '-DCYTHON_TRACE_NOGIL=1',
-        )).strip()
-    try:
-        yield
-    finally:
-        os.environ.clear()
-        os.environ.update(orig_env)
-
-
-@contextmanager
-def _run_in_temporary_directory() -> t.Iterator[Path]:
+def _in_temporary_directory(src_dir: Path) -> t.Iterator[None]:
     with TemporaryDirectory(prefix='.tmp-yarl-pep517-') as tmp_dir:
         with chdir_cm(tmp_dir):
-            yield Path(tmp_dir)
+            tmp_src_dir = Path(tmp_dir) / 'src'
+            copytree(src_dir, tmp_src_dir, symlinks=True)
+            os.chdir(tmp_src_dir)
+            yield
 
 
 @contextmanager
-def maybe_prebuild_c_extensions(  # noqa: WPS210
+def maybe_prebuild_c_extensions(
+        line_trace_cython_when_unset: bool = False,
         build_inplace: bool = False,
         config_settings: dict[str, str] | None = None,
 ) -> t.Generator[None, t.Any, t.Any]:
@@ -260,12 +218,27 @@ def maybe_prebuild_c_extensions(  # noqa: WPS210
     :param config_settings: :pep:`517` config settings mapping.
 
     """
+    cython_line_tracing_requested = _include_cython_line_tracing(
+        config_settings,
+        default=line_trace_cython_when_unset,
+    )
     is_pure_python_build = _make_pure_python(config_settings)
 
     if is_pure_python_build:
         print("*********************", file=_standard_error_stream)
         print("* Pure Python build *", file=_standard_error_stream)
         print("*********************", file=_standard_error_stream)
+
+        if cython_line_tracing_requested:
+            _warn_that(
+                f'The `{CYTHON_TRACING_CONFIG_SETTING !s}` setting requesting '
+                'Cython line tracing is set, but building C-extensions is not. '
+                'This option will not have any effect for in the pure-python '
+                'build mode.',
+                RuntimeWarning,
+                stacklevel=999,
+            )
+
         yield
         return
 
@@ -281,26 +254,15 @@ def maybe_prebuild_c_extensions(  # noqa: WPS210
             stacklevel=999,
         )
 
-    original_src_dir = Path.cwd().resolve()
     build_dir_ctx = (
-        nullcontext(original_src_dir) if build_inplace
-        else _run_in_temporary_directory()
+        nullcontext() if build_inplace
+        else _in_temporary_directory(src_dir=Path.cwd().resolve())
     )
-    with build_dir_ctx as tmp_dir:
-        if not build_inplace:
-            tmp_src_dir = Path(tmp_dir) / 'src'
-            copytree(original_src_dir, tmp_src_dir, symlinks=True)
-            os.chdir(tmp_src_dir)
-
+    with build_dir_ctx:
         config = _get_local_cython_config()
 
-        py_ver_arg = f'-{_python_version_tuple.major!s}'
-
-        cli_flags = get_enabled_cli_flags_from_config(config['flags'])
-        cli_kwargs = get_cli_kwargs_from_config(config['kwargs'])
-
-        cythonize_args = cli_flags + [py_ver_arg] + cli_kwargs + config['src']
-        with patched_env(config['env']):
+        cythonize_args = _make_cythonize_cli_args_from_config(config)
+        with _patched_cython_env(config['env'], cython_line_tracing_requested):
             _cythonize_cli_cmd(cythonize_args)
         with patched_distutils_cmd_install():
             with patched_dist_has_ext_modules():
@@ -323,6 +285,7 @@ def build_wheel(
 
     """
     with maybe_prebuild_c_extensions(
+            line_trace_cython_when_unset=False,
             build_inplace=False,
             config_settings=config_settings,
     ):
@@ -349,8 +312,9 @@ def build_editable(
 
     """
     with maybe_prebuild_c_extensions(
-        build_inplace=True,
-        config_settings=config_settings,
+            line_trace_cython_when_unset=True,
+            build_inplace=True,
+            config_settings=config_settings,
     ):
         return _setuptools_build_editable(
             wheel_directory=wheel_directory,